examples/python-fastapi-integration/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2 